refactor: decouple main function to allow custom distributions of the telegraf agent
