+++ cmd/telegraf/app/printer.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	_ "embed"
